refactor: simplify dict_utils
